+++ src/main/java/edu/ie3/datamodel/io/factory/input/ThermalHouseInputFactory.java
@@ -20,7 +20,7 @@ public class ThermalHouseInputFactory
     extends AssetInputEntityFactory<ThermalHouseInput, ThermalUnitInputEntityData> {
   private static final String ETH_LOSSES = "ethlosses";
   private static final String ETH_CAPA = "ethcapa";
-  private static final String DESIRED_TEMPERATURE = "desiredTemperature";
+  private static final String TARGET_TEMPERATURE = "targetTemperature";
   private static final String UPPER_TEMPERATURE_LIMIT = "upperTemperatureLimit";
   private static final String LOWER_TEMPERATURE_LIMIT = "lowerTemperatureLimit";